Separate the database updater and flask app.
